pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-neuropitch_env"
version = "0.1.0"
description = "Neuropitch Env environment for OpenEnv"
requires-python = ">=3.11"
dependencies = [
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    # install from github
    # "openenv-core[core] @ git+https://github.com/meta-pytorch/OpenEnv.git",
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.8.0",
    "openai>=1.40.0",
    "requests>=2.32.0",
    "ollama>=0.3.0",
    "numpy>=1.26.0",
    "datasets>=2.20.0",
    "transformers==4.56.2",
    "trl==0.22.2",
    "peft<0.18.0",
    "torchao<0.12.0",
    "unsloth>=2025.3.0",
    "tribev2[plotting] @ git+https://github.com/facebookresearch/tribev2.git",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m neuropitch_env.server.app
server = "neuropitch_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["neuropitch_env", "neuropitch_env.server"]
package-dir = { "neuropitch_env" = ".", "neuropitch_env.server" = "server" }